+++ src/StrongTypes.OpenApi.TestApi.Shared/AnnotatedRequests.cs
@@ -39,27 +39,3 @@ public sealed record AnnotatedNumbersRequest(
 
 public sealed record AnnotatedTagsRequest(
     [property: MaxLength(10)] NonEmptyEnumerable<NonEmptyString> Tags);
-
-[ApiController]
-[Route("annotated-texts")]
-public sealed class AnnotatedTextsController : ControllerBase
-{
-    [HttpPost]
-    public IActionResult Echo(AnnotatedTextsRequest request) => Ok(request);
-}
-
-[ApiController]
-[Route("annotated-numbers")]
-public sealed class AnnotatedNumbersController : ControllerBase
-{
-    [HttpPost]
-    public IActionResult Echo(AnnotatedNumbersRequest request) => Ok(request);
-}
-
-[ApiController]
-[Route("annotated-tags")]
-public sealed class AnnotatedTagsController : ControllerBase
-{
-    [HttpPost]
-    public IActionResult Echo(AnnotatedTagsRequest request) => Ok(request);
-}
