+++ src/StrongTypes.OpenApi.TestApi.Shared/BasicControllers.cs
